Modified HTMLView to use RenderContext instead of CoreQuery to update HTML.
This works better because the first time render() is called, there is no layer (DOM node) for CoreQuery to manipulate! (The eventual contents of the DOM node are being buffered in the RenderContext.) So the html was only being added to the DOM as a side effect of the updating of the html property. But in tab panes (except for the default pane) the html property is actually updated *before* the layer is created, so nothing at all was showing. (Sproutcore won't generate the layer until the view isVisible)