Started to refector
Trying to make a QuestionView class that can be used across the question tabs. Some more work needs to be done to (sort-term) hide the inputView text field on some questions and (long-term) dynamically add a child view of the inputView text field when needed to an instance of QuestionView.