Updated DialogTurn fixtures to correct usage.
'isLastTurn' should be YES for the 'final' dialog turn on a given *page*, e.g., it should be YES for turn-2-correct and turn-2-incorrect.
Progression to the third sequence of dialog turns should occur by creating a third GuidePage object, 'page-3', with a firstDialogTurn property of 'turn-3'. (that's what I did here.)

Also removed reference to since-removed NoSelection verifier.