Arrow now takes a LabelUnitKey argument.
If that argument is set, and the named attribute exists for the axes of the relevant graph, the value of that attribute will be added to the arrow as "(∆%s)". I've set the createRiseArrow and createRunArrow methods to use yLabelAbbreviated and xLabelAbbreviated, respectively, but if another attribute is created on the axes at a later date, tweaking the key will change what the arrow label uses.