Change label "Select All" behavior.
When a label is first created, and its editable, its contents should all be selected.

Subsequent keystrokes replace the selected text.

When the label is being edited, and it receives a double click, the label selects its contents.
When command-a or ctrl-a is pressed, the labels contents are also all selected.

Subsequent keystrokes replace the selected text.